debian: fix issues when moving shim/grub outside of LAT
When we move shim/grub images out of lat sdk and use the packages
built by starlingx debian, issues occur. Below changes are done
to solve the errors reported:
(1)When grub.cfg's path in grub-efi-amd64*.deb is same with the path
it is to be copied as "BOOT_GRUB_CFG", this copying should be avoided;
(2)There are grub mods installed in grub-efi-amd64*.deb in path
/boot/efi/EFI/BOOT/x86_64-efi, so flag "-r" is needed when rootfs_efi
is copied;
(3)When EFI_SECURE_BOOT is disabled, bootx64-nosig.efi is copied to
bootx64.efi and there is no need to check if bootx64.efi exits or
not.

+++ recipes-support/genimage/files/genimage/genimage.py
@@ -853,7 +853,8 @@ class GenExtDebImage(GenImage):
         utils.mkdirhier(rootfs_efi)
 
         # Copy grub.cfg to rootfs
-        utils.run_cmd_oneshot("cp -f %s %s/" % (self.data['gpg']['grub']['BOOT_GRUB_CFG'], rootfs_efi))
+        if not os.path.exists(os.path.join(rootfs_efi, "grub.cfg")):
+            utils.run_cmd_oneshot("cp -f %s %s/" % (self.data['gpg']['grub']['BOOT_GRUB_CFG'], rootfs_efi))
 
         # Set environment OSTREE_MULTIPLE_KERNELS and OSTREE_DEFAULT_KERNEL for run.do_image_ostree
         kernels = list()
@@ -913,8 +914,7 @@ class GenExtDebImage(GenImage):
             # Copy no secure boot loader to rootfs if it is not available
             src = self.data['gpg']['grub']['BOOT_NOSIG_GRUB']
             dst = os.path.join(rootfs_efi, "bootx64.efi")
-            if not os.path.exists(dst):
-                utils.run_cmd_oneshot("cp -f %s %s" % (src, dst))
+            utils.run_cmd_oneshot("cp -f %s %s" % (src, dst))
 
         # Make sure deploy dir clean
         utils.run_cmd_oneshot("rm -f vmlinuz-*-amd64* *.efi* grub.cfg*", cwd=self.deploydir)
@@ -930,7 +930,7 @@ class GenExtDebImage(GenImage):
         utils.run_cmd_oneshot("cp %s/boot/vmlinuz-*-amd64* %s" % (rootfs.target_rootfs, self.deploydir))
 
         # Copy boot loader to deploy dir
-        utils.run_cmd_oneshot("cp -f %s/* %s" % (rootfs_efi, self.deploydir))
+        utils.run_cmd_oneshot("cp -r -f %s/* %s" % (rootfs_efi, self.deploydir))
 
         # Create symlink bzIamge to kernel
         for kernel in glob.glob(os.path.join(self.deploydir, 'vmlinuz-*-amd64')):
